+++ core/src/main/java/cucumber/api/cli/Main.java
@@ -22,7 +22,6 @@ public class Main {
         ResourceLoader resourceLoader = new MultiLoader(classLoader);
         ClassFinder classFinder = new ResourceLoaderClassFinder(resourceLoader, classLoader);
         Runtime runtime = new Runtime(resourceLoader, classFinder, classLoader, runtimeOptions);
-        runtime.writeStepdefsJson();
         runtime.run();
         System.exit(runtime.exitStatus());
     }