+++ examples/java-calculator/src/test/java/cucumber/examples/java/calculator/RunCukesTest.java
@@ -5,6 +5,6 @@ import cucumber.api.junit.Cucumber;
 import org.junit.runner.RunWith;
 
 @RunWith(Cucumber.class)
-@CucumberOptions(format = "json:target/cucumber-report.json")
+@CucumberOptions(format = "json:target/cucumber-report.json", dotcucumber = ".cucumber")
 public class RunCukesTest {
 }