+++ java/src/test/java/cucumber/runtime/java/JavaBackendTest.java
@@ -5,12 +5,12 @@ import cucumber.runtime.Glue;
 import cucumber.runtime.HookDefinition;
 import cucumber.runtime.StepDefinition;
 import cucumber.runtime.StepDefinitionMatch;
+import cucumber.runtime.io.ResourceLoader;
 import cucumber.runtime.java.stepdefs.Stepdefs;
 import gherkin.I18n;
 import gherkin.formatter.model.Step;
 import org.junit.Test;
 
-import java.io.IOException;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
@@ -81,7 +81,7 @@ public class JavaBackendTest {
         }
 
         @Override
-        public void writeStepdefsJson(List<String> featurePaths, URL dotCucumber) throws IOException {
+        public void writeStepdefsJson(ResourceLoader resourceLoader, List<String> featurePaths, URL dotCucumber) {
             throw new UnsupportedOperationException();
         }
     }