+++ testng/src/main/java/cucumber/api/testng/AbstractTestNGCucumberTests.java
@@ -14,29 +14,28 @@ import org.testng.IHookable;
 import org.testng.ITestResult;
 import org.testng.annotations.Test;
 
+import java.io.IOException;
+
 public abstract class AbstractTestNGCucumberTests implements IHookable {
-    private final TestNgReporter reporter;
     private final Runtime runtime;
-    private final ClassLoader classLoader;
-    private final ResourceLoader resourceLoader;
 
     public AbstractTestNGCucumberTests() {
-        classLoader = getClass().getClassLoader();
-        resourceLoader = new MultiLoader(classLoader);
+        ClassLoader classLoader = getClass().getClassLoader();
+        ResourceLoader resourceLoader = new MultiLoader(classLoader);
 
         RuntimeOptionsFactory runtimeOptionsFactory = new RuntimeOptionsFactory(getClass(), new Class[]{CucumberOptions.class});
         RuntimeOptions runtimeOptions = runtimeOptionsFactory.create();
 
-        reporter = new TestNgReporter(System.out);
+        TestNgReporter reporter = new TestNgReporter(System.out);
         runtimeOptions.getFormatters().add(reporter);
         ClassFinder classFinder = new ResourceLoaderClassFinder(resourceLoader, classLoader);
         runtime = new Runtime(resourceLoader, classFinder, classLoader, runtimeOptions);
     }
 
     @Test(groups = "cucumber", description = "Runs Cucumber Features")
-    public void run_cukes() {
+    public void run_cukes() throws IOException {
         runtime.run();
-
+        runtime.printSummary();
         if (!runtime.getErrors().isEmpty()) {
             throw new CucumberException(runtime.getErrors().get(0));
         }